+++ nodes/Autotask/helpers/http/index.ts
@@ -1,2 +1,9 @@
-export * from './headers';
 export * from './request';
+export * from './headers';
+
+// Import and export rate tracker initialization functions
+export * from './initRateTracker';
+
+// Perform early initialization of the rate tracker
+import { initializeRateTrackerEarly } from './initRateTracker';
+initializeRateTrackerEarly();